requirements.txt
Django==3.0.5
djangorestframework==3.12.4
django-cors-headers==3.7.0
PyMySQL==1.0.2